requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.8
numpy==1.24.4
pandas==2.0.3
progressbar2==4.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-utils==3.8.2
pytz==2024.1
requests==2.32.3
six==1.16.0
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
